move basic auth config into api pkg
