+++ pkg/api/basic_auth_config_errors.go
@@ -1,4 +1,4 @@
-package config
+package api
 
 type errBasicAuthUsernameNotSpecified struct{}
