+++ pkg/api/basic_auth_config_test.go
@@ -1,4 +1,4 @@
-package config
+package api
 
 import (
 	"os"
